+++ trunk/src/cz/ligas/exportoverview/gui/EditDeliveryLineForm.java
@@ -18,7 +18,7 @@ public class EditDeliveryLineForm extends EditDocumentLineForm {
     @Override
     public void saveDocumentLine() {
         int sold = 0;
-        float price = 0;
+        double price = 0;
         try {
             if (!nAmountIn.getValue().toString().equals("")) {
                 sold = MyParser.pareseIntNumber(nAmountIn.getValue().toString());
